Extract requestToken from driver
